chore: Move settings up in hierachy
