+++ wrestling_scoreboard_client/lib/view/screens/home/more.dart
@@ -4,7 +4,7 @@ import 'package:go_router/go_router.dart';
 import 'package:wrestling_scoreboard_client/view/screens/more/about.dart';
 import 'package:wrestling_scoreboard_client/view/screens/more/imprint.dart';
 import 'package:wrestling_scoreboard_client/view/screens/more/privacy_policy.dart';
-import 'package:wrestling_scoreboard_client/view/screens/more/settings/settings.dart';
+import 'package:wrestling_scoreboard_client/view/screens/more/settings.dart';
 import 'package:wrestling_scoreboard_client/view/widgets/responsive_container.dart';
 import 'package:wrestling_scoreboard_client/view/widgets/scaffold.dart';
